PR changes requested by @ostafen: unit tests in db_test.go, and return any errors from Marshalling.
